AgenticAIs/Agent-1/requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
google-generativeai>=0.3.0
python-dotenv>=1.0.0
